Replace backslashes in paths (Windows use case)
If `pixplot --images` was run from Windows, manifest.json will have backslashes in paths.
getPath() assumes some paths start with `output/` and strips this. If a path starts instead
with `output\\`, then the result from getPath will likely have the `output` directory twice.

When the user loads index.html, the first file it attemps to read (the imagelists json file)
will give a 404 error. As a result, the page will not load if the pixplot --images command
was generated on Windows.

By replacing backslashes with forward slashes, this can be fixed. I don't know if chaining
the replace operations is the best JavaScript style -- I am new to JavaScript.